+++ configure
@@ -32,8 +32,7 @@ ninja
   .description('$command')
 
 let browserify = `browserify --extension='.jsx' --transform [ babelify \
-    --presets [ @babel/react @babel/env ] ] --transform [ stringify \
-    --extensions [.html] ] `
+    --presets [ @babel/react @babel/env ] ]`
 
 if (config === 'dev') {
   browserify += ' --debug --full-paths'
@@ -228,15 +227,16 @@ function getKitspaceYaml(folder) {
 const jsSrc = globule.find(['src/**/*.js', 'src/**/*.jsx']).
       filter((p) => !p.includes('InteractiveHtmlBom'))
 
-const jsMainTargets = jsSrc.map(function(f) {
-  const temp = ninjaBuildGen.escape(f.replace('src', 'build/.temp'))
-  f = ninjaBuildGen.escape(f)
-  ninja
-    .edge(temp)
-    .from(f)
-    .using('copy')
-  return temp
-})
+const jsMainTargets = ['build/.temp/interactive_bom/IBOM.jsx'].concat(
+  jsSrc.map(function(f) {
+    const temp = ninjaBuildGen.escape(f.replace('src', 'build/.temp'))
+    f = ninjaBuildGen.escape(f)
+    ninja
+      .edge(temp)
+      .from(f)
+      .using('copy')
+    return temp
+  }))
 
 const jsPageTargets = {}
 for (const board of boards) {
@@ -257,6 +257,15 @@ for (const board of boards) {
     })
 }
 
+const sassSrc = globule.find('src/**/*.scss')
+
+ninja
+  .edge('build/.temp/index.css')
+  .from('src/index/index.scss')
+  .need(sassSrc)
+  .assign('path', 'src/')
+  .using('sass')
+
 const ibomJs = globule.find(['src/interactive_bom/InteractiveHtmlBom/*.js'])
 
 ninja
@@ -269,14 +278,10 @@ ninja
   .from(ibomJs)
   .using('concat')
 
-const sassSrc = globule.find('src/**/*.scss')
-
 ninja
-  .edge('build/.temp/index.css')
-  .from('src/index/index.scss')
-  .need(sassSrc)
-  .assign('path', 'src/')
-  .using('sass')
+  .edge('build/.temp/interactive_bom/InteractiveHtmlBom/ibom.html')
+  .from('src/interactive_bom/InteractiveHtmlBom/ibom.html')
+  .using('copy')
 
 for (const name of ['bom-builder', '1-click-bom', 'newsletter', 'interactive_bom']) {
   // compile the sass scss into css
@@ -315,11 +320,6 @@ ninja
   .need(sassSrc)
   .assign('path', 'src/')
   .using('sass')
-
-ninja
-  .edge('build/.temp/interactive_bom/InteractiveHtmlBom/ibom.html')
-  .from('src/interactive_bom/InteractiveHtmlBom/ibom.html')
-  .using('copy')
 
 ninja
   .edge('build/index.css')
